refactor some common functionality
