+++ syftr/scripts/experiments/llm_judge_adherence.py
@@ -4,10 +4,10 @@ import json
 import time
 import typing as T
 
-from ray.job_submission import JobStatus
 from slugify import slugify
 
 from syftr.configuration import cfg
+from syftr.experiments import iter_all_job_logs
 from syftr.logger import logger
 from syftr.optimization import user_confirm_delete
 from syftr.optuna_helper import get_completed_flows, get_pareto_flows
@@ -293,20 +293,6 @@ def get_optimization_parameters():
         yield DATASETS, SEARCH_SPACE, optimization_config, evaluation
 
 
-def derived_representer(dumper, data):
-    return dumper.represent_dict({"description": data.description})
-
-
-async def iter_job_logs(job_logs: T.AsyncIterable):
-    async for lines in job_logs:
-        print(lines, end="")
-
-
-async def iter_all_job_logs(tailers: T.List[T.AsyncIterable]):
-    log_iters = [iter_job_logs(tailer) for tailer in tailers]
-    await asyncio.gather(*log_iters)
-
-
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument(
@@ -371,21 +357,6 @@ def main():
     # monitor benchmarks
     log_tailers = [client.tail_job_logs(job) for job in job_ids]
 
-    asyncio.run(iter_all_job_logs(log_tailers))
-
-
-def attach_logs(prefix: str = "<doesntmatch>", remote: bool = True):
-    cfg.ray.local = False if remote else cfg.ray.local
-    client = get_client()
-    job_details = client.list_jobs()
-    jobs_to_tail = [
-        job
-        for job in job_details
-        if job.submission_id is not None
-        and job.submission_id.startswith(prefix)
-        and job.status not in {JobStatus.STOPPED, JobStatus.SUCCEEDED, JobStatus.FAILED}
-    ]
-    log_tailers = [client.tail_job_logs(job.job_id) for job in jobs_to_tail]
     asyncio.run(iter_all_job_logs(log_tailers))
